feat: `LazyListState.isAtBottom` 확장 함수 공통 유틸리티로 분리
- `RankingScreen.kt`에 정의되어 있던 `isAtBottom` 확장 함수를 `com.yagubogu.ui.util` 패키지의 `LazyListStateExt.kt`로 이동
- `RankingScreen.kt`에서 공통 유틸리티 함수를 사용하도록 수정하여 코드 재사용성 향상
