+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/ranking/component/RankingScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListLayoutInfo
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -17,9 +16,7 @@ import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -39,6 +36,7 @@ import com.yagubogu.ui.ranking.model.RankingUiModel
 import com.yagubogu.ui.theme.Gray050
 import com.yagubogu.ui.theme.Gray300
 import com.yagubogu.ui.theme.White
+import com.yagubogu.ui.util.isAtBottom
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
 import org.koin.core.parameter.parametersOf
@@ -188,25 +186,6 @@ private fun RankingContent(
         }
     }
 }
-
-@Composable
-fun LazyListState.isAtBottom(buffer: Int = 2): Boolean =
-    remember(this) {
-        // derivedStateOf를 사용하여 불필요한 리컴포지션 방지
-        derivedStateOf {
-            val layoutInfo: LazyListLayoutInfo = layoutInfo
-            val totalItemsCount: Int = layoutInfo.totalItemsCount
-            val lastVisibleItemIndex: Int = layoutInfo.visibleItemsInfo.lastOrNull()?.index ?: 0
-
-            // 데이터가 없으면 false
-            if (totalItemsCount == 0) {
-                false
-            } else {
-                // (마지막으로 보이는 인덱스 + 버퍼)가 전체 개수보다 크거나 같으면 바닥에 도달한 것으로 간주
-                lastVisibleItemIndex + buffer >= totalItemsCount
-            }
-        }
-    }.value
 
 @Preview
 @Composable
